fix: refactor constructors of entities

+++ RangoUnchained/core/src/main/java/io/github/RangoUnchained/Model/Components/PositionComponent.java
@@ -2,25 +2,25 @@ package io.github.RangoUnchained.Model.Components;
 
 public class PositionComponent implements Component {
 
-    private int posX;
-    private int posY;
+    private float posX;
+    private float posY;
 
     public PositionComponent() {
     }
 
-    public int getPosX() {
+    public float getPosX() {
         return posX;
     }
 
-    public void setPosX(int posX) {
+    public void setPosX(float posX) {
         this.posX = posX;
     }
 
-    public int getPosY() {
+    public float getPosY() {
         return posY;
     }
 
-    public void setPosY(int posY) {
+    public void setPosY(float posY) {
         this.posY = posY;
     }
 }
